+++ bashlib/src/dumpjson.bash
@@ -1,0 +1,15 @@
+## ### `ocrd__dumpjson`
+## 
+## Output ocrd-tool.json.
+## 
+## Requires `$OCRD_TOOL_JSON` and `$OCRD_TOOL_NAME` to be set:
+## 
+## ```sh
+## export OCRD_TOOL_JSON=/path/to/ocrd-tool.json
+## export OCRD_TOOL_NAME=ocrd-foo-bar
+## ```
+## 
+ocrd__dumpjson () {
+    ocrd ocrd-tool "$OCRD_TOOL_JSON" tool "$OCRD_TOOL_NAME" dump
+}
+