+++ bashlib/test/args.test.sh
@@ -1,0 +1,10 @@
+export OCRD_TOOL_JSON=./test/ocrd-tool.json
+export OCRD_TOOL_NAME="ocrd-test-bashlib"
+declare -A argv
+source `ocrd bashlib filename`
+
+ocrd__parse_argv "$@"
+
+for k in "${!argv[@]}";do
+    echo "$k == ${argv[$k]}"
+done